Move cosmos struct to its own file
